split off report.test from reports.test move some tests around
