+++ src/kernel/spell.test.c
@@ -1,6 +1,7 @@
 #include <platform.h>
 
 #include <kernel/spell.h>
+#include <kernel/spellbook.h>
 
 #include <util/log.h>
 #include <util/lists.h>
